+++ lib/mk_framework/generator.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require_relative 'generator/configuration'
+require_relative 'generator/options'
+require_relative 'generator/project'
+require_relative 'generator/cli'